게시물 수정, 삭제 권한체크, 로그인체크. jsp include 사용법

+++ src/main/webapp/jsp/article/detail.jsp
@@ -5,7 +5,7 @@
 
 <%
 	Map<String, Object> articleRow = (Map) request.getAttribute("articleRow");
-
+	int loginedMemberId = (int) request.getAttribute("loginedMemberId");
 %>
 
 
@@ -28,8 +28,13 @@
 	</ul>
 	<div>
 		<a href="list">목록</a>
-		<a href="modify?id=<%= (int) articleRow.get("id") %>">수정</a>
-		<a href="dodelete?id=<%= (int) articleRow.get("id") %>" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;">삭제</a>
+		
+		<%if (loginedMemberId != -1 && (int) articleRow.get("memberId") == loginedMemberId) { %>
+			<a href="modify?id=<%= (int) articleRow.get("id") %>">수정</a>
+			<a href="dodelete?id=<%= (int) articleRow.get("id") %>" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;">삭제</a>
+		
+		<% } %>
+		
 	</div>
 	
 </body>
